+++ lib/rubocop/cop/performance/range_include.rb
@@ -3,18 +3,19 @@
 module RuboCop
   module Cop
     module Performance
-      # This cop identifies uses of `Range#include?`, which iterates over each
+      # This cop identifies uses of `Range#include?` and `Range#member?`, which iterates over each
       # item in a `Range` to see if a specified item is there. In contrast,
       # `Range#cover?` simply compares the target item with the beginning and
       # end points of the `Range`. In a great majority of cases, this is what
       # is wanted.
       #
-      # This cop is `Safe: false` by default because `Range#include?` and
+      # This cop is `Safe: false` by default because `Range#include?` (or `Range#member?`) and
       # `Range#cover?` are not equivalent behaviour.
       #
       # @example
       #   # bad
       #   ('a'..'z').include?('b') # => true
+      #   ('a'..'z').member?('b')  # => true
       #
       #   # good
       #   ('a'..'z').cover?('b') # => true
@@ -24,7 +25,7 @@ module RuboCop
       #
       #   ('a'..'z').cover?('yellow') # => true
       class RangeInclude < Cop
-        MSG = 'Use `Range#cover?` instead of `Range#include?`.'
+        MSG = 'Use `Range#cover?` instead of `Range#%<bad_method>s`.'
 
         # TODO: If we traced out assignments of variables to their uses, we
         # might pick up on a few more instances of this issue
@@ -32,13 +33,14 @@ module RuboCop
         # (We don't even catch it if the Range is in double parens)
 
         def_node_matcher :range_include, <<~PATTERN
-          (send {irange erange (begin {irange erange})} :include? ...)
+          (send {irange erange (begin {irange erange})} ${:include? :member?} ...)
         PATTERN
 
         def on_send(node)
-          return unless range_include(node)
-
-          add_offense(node, location: :selector)
+          range_include(node) do |bad_method|
+            message = format(MSG, bad_method: bad_method)
+            add_offense(node, location: :selector, message: message)
+          end
         end
 
         def autocorrect(node)